Level 1 Template Complete
Level 1's Template and how the level will likly go for around. This goes into new Tilemaps and level size and some hints of the next level at the end of the first with a fight room.

Need 3 more tiles to complete the levels ground work and and background for further polishing.